+++ lib/tescon/trace.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
+require_relative "trace/config"
 require_relative "trace/provenance"
+require_relative "trace/attribute_normalizer"
+require_relative "trace/path_normalizer"
 require_relative "trace/recorder"
 require_relative "trace/yaml_writer"
 require_relative "trace/factory_bot"
@@ -13,6 +16,10 @@ module Tescon
     class << self
       def recorder
         @recorder ||= Recorder.new
+      end
+
+      def config
+        @config ||= Config.new
       end
 
       def reset!